pong game updated
