--- Pong-Game/ball.py
@@ -1,28 +1,0 @@
-from turtle import Turtle
-import time
-class Ball(Turtle):
-    def __init__(self):
-        super().__init__()
-        self.shape('circle')
-        self.color('white')
-        self.penup()
-        self.x_move = 10
-        self.y_move = 10
-        self.move_speed = 0.09
-    
-    def move(self):
-        new_x = self.xcor() + self.x_move
-        new_y = self.ycor() + self.y_move
-        self.goto(new_x,new_y)
-    
-    def bounce_y(self):
-        self.y_move*=-1
-
-    def bounce_x(self):
-        self.x_move*=-1
-        self.move_speed *= 0.8
-    
-    def reset_position(self):
-        self.goto(0,0)
-        self.move_speed = 0.09
-        self.bounce_x()
