--- Pong-Game/score.py
@@ -1,31 +1,0 @@
-from turtle import Turtle
-ALIGNMENT="center"
-FONT=("Courier", 80, "normal")
-class Score(Turtle):
-    def __init__(self):
-        super().__init__()
-        self.color('white')
-        self.penup()
-        self.hideturtle()
-        self.l_score = 0
-        self.r_score = 0
-        self.update_score()
-    
-    def update_score(self):
-        self.clear()
-        self.goto(-100,200)
-        self.write(self.l_score,align=ALIGNMENT,font=FONT)
-        self.goto(100,200)
-        self.write(self.r_score,align=ALIGNMENT,font=FONT)
-        
-    def l_point(self):
-        self.l_score += 1
-        self.update_score()
-    
-    def r_point(self):
-        self.r_score += 1
-        self.update_score()
-    
-    def game_over(self):
-        self.goto(0,0)
-        self.write("Game Over",align=ALIGNMENT,font=FONT)
